requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==0.28.1
langchain==0.0.335
chromadb==0.4.22
sentence-transformers==2.2.2
python-dotenv==1.0.0
PyPDF2==3.0.1
PyMuPDF==1.23.8
pathlib2==2.3.7.post1